Add a SYMBOLS layer
It basically maps the symbols from the number keys down to the home row.
Also maps the numbers to one row above home and the F keys to the number row.
I still have some issues to figure out about which modifiers to use to access the layers
but this is getting closer to finished.